Make static v. page logic more robust
Previously this checked if a file existed and then read it to serve a
static file or a page. This switches to just reading the file and
dealing with the error, which prevents a race condition where the file
is deleted before it can be read.

This also cleans up the path cleaning logic, though it’s unclear to me
whether it’s necessary.