New architecture: actions and `Return`s.
This sets up the transition to configurable rules. It adds `Return`s
for values returned by “actions”, e.g. read a path, or convert render in
a template.

There are no longer `Page`s; instead there is a function to convert a
markdown page file (with metadata) into a generic `ContentReturn`
containing the metadata and HTML. That object can be passed to the
`render()` function to render it into a template.

Most tests are passing; currently redirecting to the canonical path
doesn’t work.